d3: suprisingly terse and dependent free version of clocky with d3
First made a native version of the clocky program in
clocky-js.html. Then tried a very straightforward
port of the code to cljs, and it worked well.

Note that the cljs version is passing around native clojure
maps, not java objects. The only caveat is that these need
to be wrapped in a native java array, but that seems a small
price to pay for interop with otherwise fairly idomatic
cloure data structures.