Fixed clocky rotation bug in d3 based versions.
Thanks to Mark David at the meetup for pointing out a
visual bug in the last couple of versions in my slideshow.
Turns out that c2 and d3 have different ideas of what
an arc "startAngle" should be. c2 is at 3 o'clock
and d3 is at 12 o'clock.

So all that was needed was to drop the 90 degree
rotation transform in the d3 versions.