strokes: first working example of metaprogramming with d3
In the hull example, the d3.geom.hull method is given an
array of patched cljs vectors with minimally metaprogrammed
accessors for the elements 0 and 1, as well as a length
method and an updated toString method. d3.geom.hull used
this to unwittingly generate a new array of cljs vectors
which are then passed off to be drawn.

The same array of patched cljs vectors is given to the
.data call to draw the circles directly. In both cases,
the objects are accessible as read-only elements and
generate the correct string representation to generate
the svg needed.

Now will generalize the patch-vector call and look into
making a similar method for the top level seq to also
eliminate the (apply array) call.