mrhyde: patching all clojure datatypes at once!
By patching the persistent vector prototype (__proto__
member actually), we magically can get rid of all
strokes/patch-vector calls. This is pretty huge.
The 7 lines of magic in fact will be repeated here:

  (let [v [1, 2]
        p (.-__proto__ v)]

    (.__defineGetter__ p "length" #(this-as t (count t)))
    (dotimes [n 100]
      (.__defineGetter__ p n #(this-as t (nth t n js/undefined))))
    (-> p .-toString (set! #(this-as t (clojure.string/join ", " t)))))

This is run at launch time and effectively patches all persistent
vector type to pretend to be arrays good enough for the d3 interop
needed for the hull example.

Ideally, the getters for the individual elements would be installed
based on the length. This should be possible in the future by
intercepting the constructor, but for now this 100 loop kludge
serves fine.

Now let's try the same thing on the seq. Then finally, we can
try to patch the return value of d3.mouse so that all the
interop noise will be gone.