mrhyde: now patching several seq types with js array interface
With latest patching on seqs, we no longer need to (apply array foo).
For the hull example, this means that the seq of vertices can
be passed directly to d3.

Ironically, this broke the original d3.data patch somehow,
but that's just as well because it is no longer needed.
However, we can still try to patch the library return values,
which we will do next.