quadtree example now working - added transients
Switched from native arrays to transients to keep the
speed without using native javascript arrays. The nodes
method now builds a transient vector of nodes in the
quadtree, and returns the persistent version. And the
brushed function internally uses mutable tranient sets
to determine membership in the scanned and selected
groups, and updates the svg classes appropiately.

Sadly, the point and quadtree reference variables
had to become global for the brush callback to work,
but that's probably fine.

Lastly, mrhyde now provides arrayish classes a map
call, which is almost exactly like the preexisting
forEach call.