mrhyde: Abstracted patching js return values into cljs
There is now a dedicated function patch-fn1-return-value
which will wrap an existing js function so that the return
value has js->cljs applied. This is currently being tested
with d3.mouse, with a slight wrinkle. Because d3.mouse is
also used internally by d3, and it sometimes (in the brush
code) loppily mutates the return value, this can't be done
all the time. So for now there is a strokes call
(patch-mouse) to enable this behavior.