mrhyde: abstracted patching js method arg keywords into functions
Now a new function in mrhyde patch-args-keyword-to-fn2 which
scans all arguments to the provided javascript method and
replaces any keywords with a function that pulls the key
from a map. [ ie: replaces :foo with #(foo %) ] This is currently
being installed on the d3.selection.prototype.attr call
with great success.

Also cleaned up the ever shrinking strokes/core.cljs file.