mrhyde: successfully patched PersistentVector initialization
In the most cavalier (foolish?) move yet, I replace
the cljs.core.PersistentVector object with my own
wrapper. Everything works, but this technique not
surprisingly doesn't help other types like LazySeq.
Further thought needed to think about next steps.

For now, committing this patch in hopes that it
is useful for map types, which are next.