blade: initial mrhyde support for leaflet.js
Began project to take mrhyde scaffolding and apply it
to a second javascript library: leaflet.js. Thus far I
only have a hello-world program working called
tutorial1. Even that took a new mrhyde method:

patch-args-seq-to-array

Because deep in the leaflet code we find this
unfortunate function:

L.latLng = function (a, b) { // (LatLng) or ([Number, Number]) or (Number, Number)
  if (a instanceof L.LatLng) {
    return a;
  }
  if (a instanceof Array) {
    return new L.LatLng(a[0], a[1]);
  }
  if (isNaN(a)) {
    return a;
  }
  return new L.LatLng(a, b);
};

Well, instanceof Array is not going to happen, but it's
easy enough to patch that one.

Also a very ugly workaround for a new issue. The strokes
library and examples all assume that there is a d3 in
the global environment, and when this is not the case
the (def)s freak out. So the ugly hack for now is
to wrap everything in a big if/do to first check
for the existence of the core library. This is all
of course an artifact of having everything compile
down to one js file, which will not be the case
forever.