blade: completed first leaflet example
Oops, I had left out the polygon and popup from the example
earlier. Now this has been added.

Also patch-js-with-key-lookup adds clojurescript lookup by
keyword to all javascript objects. This code is a lightly
debugged version of a snippet I found in kevin's code:

https://github.com/lynaghk/todoFRP

I'm a little worried that mixing this into my existing mrhyde
abstractions might have some unintended consequences, but I can
not resist the ability to destructure js objects. :) I mean, look
at this:

    (let [popup (-> L .popup)]
      (.on mappy "click" (fn [{:keys [latlng]} e]
      (-> popup (.setLatLng latlng)
                (.setContent (str "You clicked the map at " latlng))
                (.openOn mappy)))))))

Lovely. So we'll try that on like a new jacket for a few days and
hope we don't have to return it back to the store.