all: renaming and refactoring launch patterns
Inverted logic of launching demos - now simply define a variable
before including strokes.js. This is a little cheesy, but the
idea of packing the demos in the same js file is already suspect.

Also renamed the html pages so that it was clear which ones
go with wich library.

Finally, death to core. Why does the clojure community love
core so much? It's like websites the insist on putting the www
in front of their name - it's a completely unnecessary
nesting and level of indirection that needs to get got.