added test framework
Finally, a test framework. This one - for now - stolen from
domina. Thanks Luke!

I adapted the domina test suite to be able to easily run
tests from different libraries (strokes, mrhyde, etc).
For now, this addes domina as a dependency.

I have my own ideas for replacing this in the future with
a test suite that uses strokes for layout instead and shows
test dependencies in an interesting way with a bubble layout,
but that can certainly wait for another day. Like when mutable
mrhyde data structures work.