mrhyde: now implements mutable js-cache on seq types
sequential types in clojurescript are now patched with
more sophisticaed machinery to deal with attempted mutations.
The setter will lazily instantiate a transient to refelect the
accumulated changes attempted on this object. The getter will
return either the value from this 'dirty cache' or the original
persistent value.

In addition, a protocol was defined so that cljs types can
respond to queries:

 hyde? - is this persistent collection actually a 'hyde' type
 has-cache? - has there been attempted mutations from the js world
 from-cache - render a copy of this collection with js mutations

This protocol was added to the same set of known sequential types.

Also a unit test was added to mrhyde unit tests to verify the
desired result upon mutation attempts. There is obviously much
more work to do to support all varieties of mutation, but this
takes care of the most straightforward setter case.

Also cleaned up some mrhyde code introducing if-not.