strokes: circle-pack using clj data structures and keyword access
There is a lot of data being written 'to the cache' on the
objects, and that is still being accessed with aget. But the
persistent subset is now being fetched via keyword access,
including a keyword function being passed directly to
d3.layout.pack.value.

Also added hyde protocol to mappish types, but the real
meat of the implementation is still missing. Once that
is done, there will need to be a new patch function which
does a deep "from-map" pass on certain functions.