mrhyde: added repersist, circle-pack now running on full dataset
The strokes demo circle-pack is now running on the full dataset
and the older demos like quadtree are also running well. The
mapping recurse-from-cache was removed from the d3.data call
and is now called as needed (because it's slow walking and
regenerating thousands of objects).

To help with this is a new mrhyde "repersist", which is meant
to convert what were formerly persistent objects once again into
real persistent objects. It has a nice feature that it will
auto-detect if the argument is a function and recurse appropriately
waiting for a non-function to convert.

Things are snappy enough where I can finally provide the
circle demo with the full dataset. It's not snappy, but
it's not completely embarassing either given the amount
of heavy lifting necessary to get it to work at all.