mrhyde: repersist now takes options for keys to skip
repersist now takes optional arguments, the first of which is
a key or list of keys to ignore when recursing.

This incidentally made strokes-circle pack run much faster: because
of the way that the data function is slicing and returning data
from the tree both children and parent should be ignored.