mrhyde/strokes: added exports. experimental typepatcher.
Wasted a few hours becasue google chrome is optimizing
away javascript semantics. Strong claim I know.

The machinery is in typepatcher to install a
hook on PersistentHashMap. This should hook into the
constructor, which looks like this in the compiled js:

  cljs.core.PersistentHashMap = function(meta, cnt, root, has_nil_QMARK_, nil_val, __hash) {
    this.meta = meta;
    this.cnt = cnt;
    this.root = root;
    this.has_nil_QMARK_ = has_nil_QMARK_;
    this.nil_val = nil_val;
    this.__hash = __hash;
    this.cljs$lang$protocol_mask$partition1$ = 4;
    this.cljs$lang$protocol_mask$partition0$ = 16123663
  };

But it doesn't work. Unless i put a console.log in the code above.
Or just a console. or just "3;". Its nuts. On safari and firefox
I get the desired behavior, but chrome seems to optimize
any funciton that consists of purely this.foo = bar, ignoring
setters in the prototype chain.

So maybe now I really give up. Shucks, it was a pretty good
idea too.