mrhyde/strokes: patch-args-map-to-obj for config map objects
d3's multi-value maps are awesome, and not widely used. See

http://bl.ocks.org/mbostock/3305515

But the implementation crawls all keys of the map and tries
to apply everything to the DOM, which doesn't work because
of all the extra cljs$ cruft.

This is a pattern seen in other libs as well, so for now I
simply add a patch call that maps this explicitly. Works
well with the voroni-clipping example in strokes-examples.
I doubt this will become a performance problem because these
"config" type maps tend to be small and flat.