mrhyde: aggressive hiding of member variables in clojure types
The clojure implementation doesn't seem to need visibility
of variables in maps, seqs, or their prototypes. So I am
now hiding most of those.

The effect should be that hyde-maps are now enumerable from
JavaScript via either Object.keys or (for k in m), because
the cljs cruft (and mrhyde cruft) is now effectively hidden -
except for the newly installed key/value pairs.